refactoring, delete .gitignore file
